Add accrued revenue
